Remove teleport concept
My idea is to check the change from point A to point B in movement. If
there is a collision in that change, then it'll use "preciseMovement"
after undoing "defaultMovement". This allows to really check what gets
collided first. However I can't seem to find the initial collision in
the first place. We can try this and if it doesn't work try something
else.